# Sign bytes