--- a/types/proposal.go
+++ b/types/proposal.go
@@ -41,32 +41,8 @@ func (p *Proposal) String() string {
 }
 
 func (p *Proposal) WriteSignBytes(chainID string, w io.Writer, n *int, err *error) {
-
-	wire.WriteJSON(
-		struct {
-			ChainID  string `json:"chain_id"`
-			Proposal struct {
-				BlockPartsHeader PartSetHeader `json:"block_parts_header"`
-				Height           int           `json:"height"`
-				POLBlockID       BlockID       `json:"pol_block_id"`
-				POLRound         int           `json:"pol_round"`
-				Round            int           `json:"round"`
-			} `json:"proposal"`
-		}{
-			chainID,
-			struct {
-				BlockPartsHeader PartSetHeader `json:"block_parts_header"`
-				Height           int           `json:"height"`
-				POLBlockID       BlockID       `json:"pol_block_id"`
-				POLRound         int           `json:"pol_round"`
-				Round            int           `json:"round"`
-			}{
-				p.BlockPartsHeader,
-				p.Height,
-				p.POLBlockID,
-				p.POLRound,
-				p.Round,
-			},
-		},
-		w, n, err)
+	wire.WriteJSON(CanonicalJSONOnceProposal{
+		ChainID:  chainID,
+		Proposal: CanonicalProposal(p),
+	}, w, n, err)
 }