Use canonical forms of structs for WriteSignBytes

https://github.com/tendermint/tendermint/issues/315