@ebuchman 2016-12-17
```
   before
   BenchmarkProposalWriteSignBytes-4         500000              2522 ns/op
   BenchmarkProposalSign-4                    10000            109518 ns/op
   BenchmarkProposalVerifySignature-4         10000            208273 ns/op
    
   after
   BenchmarkProposalWriteSignBytes-4         200000              8092 ns/op
   BenchmarkProposalSign-4                    10000            114741 ns/op
   BenchmarkProposalVerifySignature-4         10000            215933 ns/op
```

So it definitely slows down a bit, but I think the improvement in code readability more than makes up for it, especially since we will be writing a compiler for go-wire down the road. Also, didn't expect verification to take longer than signing. Is that right?