@jaekwon 2016-12-18
Agreed, totally worth it.

Yeah verification takes longer.  We haven't optimized via Ed25519 batch-verification though.